--- src/main/java/TollCalculator.java
@@ -1,73 +1,0 @@
-
-import java.text.SimpleDateFormat;
-import java.time.*;
-import java.time.temporal.ChronoUnit;
-import java.util.*;
-import java.util.concurrent.*;
-
-public class TollCalculator {
-    public static void main(String[] args) {
-        Vehicle vehicle = new Car();
-        Date date = new Date(122, 10, 7, 16, 33);
-
-        int fee = getTotalFee(vehicle, date);
-
-        System.out.println(fee);
-    }
-    /* Well, who doesn't love bonus points? The clip in the readme file is from the 1995 classic "Hackers".
-        I've actually seen the film through an ex-girlfriend when I was in high school. However
-        I must admit I didn't recognize the clip from memory, but a simple google search of the phrase
-        "Hack the planet" gave me the answer. Where would we be without google though?
-     */
-
-    /**
-     * Calculate the total toll fee for one day
-     *
-     * @param vehicle - the vehicle
-     * @param dates   - date and time of all passes on one day
-     * @return - the total toll fee for that day
-     */
-    public static int getTotalFee(Vehicle vehicle, Date... dates) {
-        // convert dates to List of LocalDateTime
-        List<LocalDateTime> passageTimes = new ArrayList<>();
-        for (Date date : dates) {
-            passageTimes.add(date.toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime());
-        }
-
-        LocalDateTime intervalStart = passageTimes.get(0);
-
-        int totalFee = 0;
-        for (LocalDateTime date : passageTimes) {
-            int nextFee = getTollFee(date, vehicle);
-            int tempFee = getTollFee(intervalStart, vehicle);
-
-            long minutes = ChronoUnit.MINUTES.between(intervalStart, date);
-
-            if (minutes <= 60) {
-                if (totalFee > 0) totalFee -= tempFee;
-                if (nextFee >= tempFee) tempFee = nextFee;
-                totalFee += tempFee;
-            } else {
-                totalFee += nextFee;
-            }
-        }
-        if (totalFee >  TollTimesAndFees.maxTotal) totalFee = TollTimesAndFees.maxTotal;
-        return totalFee;
-    }
-
-
-    public static int getTollFee(final LocalDateTime passageDateTime, Vehicle vehicle) {
-        LocalDate passageDate = passageDateTime.toLocalDate();
-        Boolean isTollFreeDate = new TollFreeDateCheck().isTollFreeDate(passageDate);
-
-        if (isTollFreeDate || vehicle.isTollFree()) return 0;
-
-        LocalTime passageTime = passageDateTime.toLocalTime();
-        int fee = new Fee().getFee(passageTime);
-
-        return fee;
-    }
-
-
-}
-
